some tests ported to OSX
